Avoid trying to write test data in production
See https://gitlab.com/oersi/oersi-etl/-/issues/46

+++ src/main/java/oersi/ETL.java
@@ -95,6 +95,9 @@ public class ETL {
     }
 
     private static void writeTestOutput(File fileOrDir) throws IOException {
+        if (!new File(OUT_FILE.getParent()).exists()) {
+            return;
+        }
         try (FileWriter w = new FileWriter(OUT_FILE)) {
             if (fileOrDir.isDirectory()) {
                 for (File json : fileOrDir.listFiles((d, f) -> f.toLowerCase().endsWith(".ndjson")
